fix(invitations): workspace test consolidation
